+++ lib/styles/ServerSideStyles.tsx
@@ -1,4 +1,5 @@
 import { GenerateId } from 'jss';
+import { minifyCss } from './minifyCss';
 import { createGenerateId, SheetsRegistry } from 'react-jss';
 import React, { ReactElement } from 'react';
 
@@ -13,8 +14,12 @@ class ServerSideStyles {
   }
 
   public generateStyleTag (): ReactElement {
+    const styles = this.registry.toString();
+
+    const minifiedStyles = minifyCss(styles);
+
     return (
-      <style id='server-side-styles' dangerouslySetInnerHTML={{ __html: this.registry.toString() }} />
+      <style id='server-side-styles' dangerouslySetInnerHTML={{ __html: minifiedStyles }} />
     );
   }
 }