+++ lib/styles/minifyCss.ts
@@ -1,0 +1,7 @@
+import { oneLine } from 'common-tags';
+
+const minifyCss = function (css: string): string {
+  return oneLine(css);
+};
+
+export { minifyCss };