+++ test/integration/components/layout/ContainerTests.ts
@@ -1,7 +1,7 @@
 import { assert } from 'assertthat';
-import { browser } from '../../shared/browser';
+import { browser } from '../../../shared/browser';
 import { Page } from 'puppeteer';
-import { getIntegrationTestUrl, integrationTestTimeOut } from '../../shared/environment';
+import { getIntegrationTestUrl, integrationTestTimeOut } from '../../../shared/environment';
 
 suite('Container', function (): void {
   this.timeout(integrationTestTimeOut);